model_development/model_serving_cloud_run/requirements.txt
gunicorn==23.0.0
google-cloud-storage==2.19.0
Flask==3.1.0
google-cloud-aiplatform==1.85.0
python-dotenv==1.0.1
flask-cors==5.0.1
pandas==2.2.3
